Move fields into own module
