+++ galaxy/api/v2/serializers/collection.py
@@ -34,37 +34,9 @@ __all__ = (
 )
 
 
-class NamespaceObjectField(serializers.Field):
-    """Return namespace object for a serializer field."""
-    def to_representation(self, value):
-        return {
-            'id': value.pk,
-            'href': reverse(
-                'api:namespace_detail',
-                kwargs={'pk': value.pk},
-                request=self.parent.context.get('request'),
-            ),
-            'name': value.name,
-        }
-
-
-class VersionDetailUrlField(serializers.Field):
-    """Return version detail url under collection namespace and name."""
-    def to_representation(self, value):
-        return reverse(
-            'api:v2:version-detail',
-            kwargs={
-                'namespace': value.collection.namespace.name,
-                'name': value.collection.name,
-                'version': value.version,
-            },
-            request=self.parent.context.get('request'),
-        )
-
-
 class VersionSummarySerializer(serializers.ModelSerializer):
     """Collection version with short summary of data."""
-    href = VersionDetailUrlField(source='*')
+    href = fields.VersionUrlField(source='*')
 
     class Meta:
         model = models.CollectionVersion
@@ -74,9 +46,9 @@ class VersionSummarySerializer(serializers.ModelSerializer):
 class VersionDetailSerializer(serializers.ModelSerializer):
     """Collection version with detailed data."""
     id = serializers.IntegerField(source='pk')
-    href = VersionDetailUrlField(source='*')
+    href = fields.VersionUrlField(source='*')
     download_url = serializers.SerializerMethodField()
-    namespace = NamespaceObjectField(source='collection.namespace')
+    namespace = fields.NamespaceObjectField(source='collection.namespace')
     collection = serializers.SerializerMethodField()
     metadata = serializers.JSONField(binary=False)
 
@@ -120,7 +92,7 @@ class VersionDetailSerializer(serializers.ModelSerializer):
 
 
 class CollectionSerializer(serializers.ModelSerializer):
-    namespace = NamespaceObjectField()
+    namespace = fields.NamespaceObjectField()
     href = serializers.SerializerMethodField()
     versions_url = serializers.SerializerMethodField()
     highest_version = VersionSummarySerializer()
@@ -170,7 +142,7 @@ class CollectionImportSerializer(BaseTaskSerializer):
     messages = _MessageSerializer(many=True)
     lint_records = serializers.JSONField()
 
-    namespace = NamespaceObjectField()
+    namespace = fields.NamespaceObjectField()
     imported_version = serializers.SerializerMethodField()
 
     class Meta:
